+++ SysytemAdmin/Admin.Model/Entitys/Role.cs
@@ -20,7 +20,7 @@ namespace Admin.Model.Entitys
         /// 顺序
         /// </summary>
         [SugarColumn(IsNullable = false)]
-        public long Order { get; set; }
+        public int Order { get; set; }
        
         /// <summary>
         /// 是否启用（0=未启用，1=启用）
